Adjust unit_state metric and add changelog
The next commit will add a new metric to re-add the type
field, but for now this commit makes unit_state work for
all units instead of only working for service or mount type
units. Instead of just making other unit varieties return an
empty type field, I chose to remove the type field from the
metric and add it back in the next commit. To ensure this commit
is self-contained, it also comments out the contents of
collectServiceState and collectMountState in favor of allowing
collectUnitState to handle everything