Refactor, no functional changes
Collector is a confusing name since it's overloaded - the prometheus
library calls the API we are using a "Collector". Before this change, we
were building a custom "Collector" type that also happened to be named
"Collector". This just renames it to anything else so we can refer to them
distinctly