Read CPU usage from cgroups
An initial patch to read CPU usage for the entire unit by reading
directly from the CPU hierarchy. Currently assumes v1 cgroup API
and a specific systemd version. Will improve later to be more flexible.

Through manual comparison with top and systemd_process_cpu_seconds_total
this appears to be consistent for most processes, but much more accurate for
units that spawn hundreds of threads. This is to be expected, the main failing of
the _process_ unit is that it only tracks the main process.

This patch is a bit bulky at the moment because I am intentionally attempting to
have a clean separation between the "read cgroup assuming systemd conventions"
logic and the "use dbus to acquire cgroup name" logic. I'm suspecting we could
upstream the cgroup logic once we have a solid prototype here, and then benefit
from the entire go-systemd community helping us keep the read-cgroup logic
updated. This would be a big win given that the kernel is currently undergoing a
transformation to replace the cgroupv1 with a v2 API, so it's almost certain we
will make a mistake somewhere or not test on all system configurations