Add type label for all unit_* metrics
For all metrics named unit_* (e.g. unit_state, unit_info, etc) we add a label
type that exports the unit type e.g. "service", "mount", "scope", etc. To
ensure current dashboards keep working, we do not change the contents
of the name label. Queries that rely on the "name.type" format continue
working, but for all unit_* labels we are sharing the type in two places now
(the second half of the name label, and the type label). It's inefficient, so
would be nice to update the name label when we are ready to introduce
breaking changes.

Without a distinct type label, we were unable to easily utilize grouping in
PromQL because it is impossible to group on part of a label. For example,
this is impossible:
    count(systemd_unit_state) by (name=~".*service")
while this is possible:
    count(systemd_unit_state) by (type)

According to Brian Brazil (Prometheus developer), "combined labels" like
our name label are the wrong  way to use prometheus and should be split.
If you cannot split in the exporter, the next best is to use metric_relabel_config