Fix: Stop assigning CPU time to inactive units
Noticed a bug where some units were getting assigned suspiciously high
amounts of CPU utilization. Turns out systemd returns an empty string
for the ControlGroup when there is no cgroup for that unit. However the
empty string, when concatenated with the base path for the cpu hierarchy,
corresponds to an actual cpuacct.usage_all file for the init.scope. So allowing
the read does not result in a hard error, it just results in us recording
inaccurate data for this unit at this time.

In all cases I found this was because the unit was currently inactive, so we just
ignore this situation as an expected situation, while also refusing to call the
sub-method in other cases. According to systemd code, if the intended target
is -.scope (the root scope), then the ControlGroup will be "/" allowing a valid
path combination. So this patch should work for both the root and for all non-
root units. Yay!