Stop exporting CgroupUnified & rename
There is little reason right now to export this data, and exporting it means we are tied
to an annoying naming scheme. Better to indicate this is internal only for the moment,
but leave the documentation for anyone reading.

This also stops returning a pointer to a uint8, and checks for error on internal calls