Stop spamming logs re: socket cpu accounting
Not sure how to handle this one. While almost all socket units have a ControlGroup set, almost none of those cgroups actually exist in the filesystem. Unfortunately, the ones that do exist are for critical services such as docker, which we would be foolish not to track. Reading CPUAccounting does not help, most sockets have this set to True but the on-disk control group does not exist. On my system, we have 19 of these warnings being printed with each scrape.