filterUnits now returns channel
This prevents the ~15ms delay while we run a regex over each and every
unit name. As soon as we are interested in a unit, we send it forwards in
the processing.

Unfortunately the effectiveness of this is limited by how we currently
retrieve units as a large slice all at once. If we could have a non-blocking api
for that we could go faster