Allow pprof on live system
This is also done in node_exporter, and a few other high-profile
packages. Allows you to send HTTP requests to gather tracing data
from the server, e.g. :
    go tool pprof http://<your_ip>:9558/debug/pprof/heap
OR
    curl http://<your_ip>:9558/debug/pprof/trace?seconds=15 > trace.out
    go tool trace trace.out

Overhead is effectively noise until someone calls the method to download
a trace, at which point overhead is dependent on the type of trace being
downloaded.