Initial test script based on node_exporter
Main problem is that this does not support test coverage, because we are using an
external binary as the binary-under-test and coverage works by compiling a custom
binary with coverage-tracking-logic when running go test. It's not simple to force
compilation of these tracking statements, so as long as we use an external binary we
cannot have coverage data.

Minor issue is that the process launch time is relatively slow. Not an issue when
running 2-3 tests, but this is may be an issue down the road