Allow http.Shutdown call for testing
This is a minimal proof-of-concept for how to test the exporter while
also getting code coverage details. Should also be a bit faster.

Bad things: If something ever goes wrong, cleaning up would be
much harder because we cannot just kill the external process. So
arguably this could be a bad thing, or in the future we bother to split
up testing into smoke tests that can run in this manner and larger
tests that should be run as an external process. But for now I'm happy
with this, I like seeing code coverage reports :-)

While it's tempting to think this magically enables parallelism in the
test suite, remember we are sharing the global default http server here.
It would not be too hard to change that, and also auto-update the port
numbers, but it's overkill for now. I'll just run stuff serially until it
becomes an issue