Stop trying to use circleCI
Their focus on testing-inside-docker means they do not have a good
testing-in-vm experience. Only one ubuntu version is supported, and
upgrading it on each CI run is both wasteful and a huge pain to get
correct (we only want to upgrade golang, but they have a customized
version of ubuntu to the typical backported PPA approach does not work
out of the box). As a nice side bonus, travis-ci is mostly OSS while
circleCI is not, so we can read the source if we have any issues with travis.