Add tests for complex 'systemd mount mode' logic
This exposes the critical statfs call used inside cgUnifiedCache as a
settable variable. Using this variable, we can now write test tables
that emulate common mounting scenarios. This achieves 100% coverage
within the tricky function, so we should now be able to rewrite this function in
canonical go with much less concern that we've borked it up.

Also caused me to notice+fix some bugs in the cgUnifiedCached method - we
were using file paths when we should have been explicitly using statfs on
directory paths