Add timer next trigger
This metric shows next time a timer will fire.
Care has been taken to handle real-time and monotonic timers correctly.

Real-time timers return their value directly as a Unix timestamp as microseconds.
The only special case is when the corresponding service is running,
the timestamp is the max unit64 value. We filter those values out.

Monotonic timers return the next time a service will run
as the number of microseconds since the system boot time.
Some computation is needed to get the next time as a standard Unix timestamp.

Signed-off-by: ibizaman <ibizaman@tiserbox.com>